v1.241
Implemented online pause feature.
Slightly modified UI Colors.
Fixed a case where you would freeze on drift select in time trials after playing a OTT race.
Fixed a case where repick protection wouldn't reset in a new friend room.
Fixed a case where Names Tags on Countdown would disappear forever.
Attempted to fix a case where VR would reset due to starting a worldwide in a friend room.
VR now saves directly on your SD Card in a separate file.
Changed save location of OptPack to prevent VR resetting to 10K after playing Retro Rewind.
With this change, this will be making a copy of the Wii's Save. Please check #faq on how to copy your RR save into the new save location.

+++ Pulsar/GameSource/core/nw4r/lyt/Window.hpp
@@ -36,7 +36,7 @@ public:
     const ut::detail::RuntimeTypeInfo* GetRuntimeTypeInfo() const override; //0xc 8007f2f0
     void DrawSelf(const DrawInfo& info) override; //0x18 8007d1b0
     virtual void AnimateSelf(u64 option); //0x20 8007d2f0
-    ut::Color GetVtxColor(u64 id) const override; //0x24 8007d110
+    ut::Color GetVtxColor(u32 id) const override; //0x24 8007d110
     void SetVtxColor(u64 id, ut::Color color) override; //0x28 8007d140
     u8 GetVtxColorElement(u64 id) const override; //0x34 8007d170
     void SetVtxColorElement(u64 id, u8 value) override; //0x38 8007d190
